punto8

+++ src/main/java/edu/eci/pdsw/sampleprj/dao/ClienteDAO.java
@@ -12,7 +12,7 @@ public interface ClienteDAO {
 	
 	 public void save(Cliente it) throws PersistenceException;
 
-	 public Cliente load(int id) throws PersistenceException;
+	 public Cliente load(long id) throws PersistenceException;
 	 
 	 public List<Cliente> consultarClientes() throws PersistenceException;
 }
